requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pyjwt[crypto]==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
itsdangerous==2.1.2
